Permissionamento de psicólogos

+++ app/Http/Controllers/PsychologistsController.php
@@ -3,9 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\PsychologistRequest;
+use App\Models\Psychologist;
 use App\Repositories\PsychologistsRepository;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\HttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class PsychologistsController extends Controller
@@ -19,6 +22,8 @@ class PsychologistsController extends Controller
 	 */
 	public function index(Request $request)
 	{
+		$this->authorize('index', Psychologist::class);
+
 		return $this->repository->getAll(collect($request->all()));
 	}
 
@@ -27,6 +32,8 @@ class PsychologistsController extends Controller
 	 */
 	public function store(PsychologistRequest $request)
 	{
+		$this->authorize('store', Psychologist::class);
+
 		$user = $this->repository->create($request->all());
 
 		event(new Registered($user));
@@ -39,10 +46,12 @@ class PsychologistsController extends Controller
 	 */
 	public function edit($id)
 	{
+		$this->authorize('edit', Psychologist::class);
+
 		$user = $this->repository->getById((int) $id);
 	
 		if (!$user) {
-			throw new NotFoundHttpException("Nenhum registro encontrado");
+			throw new NotFoundHttpException("Nenhum registro encontrado.");
 		}
 	
 		return $user;
@@ -53,10 +62,12 @@ class PsychologistsController extends Controller
 	 */
 	public function update(PsychologistRequest $request, $id)
 	{
+		$this->authorize('update', Psychologist::class);
+
 		$user = $this->repository->update((int) $id, $request->validated());
 
 		if (!$user) {
-			throw new NotFoundHttpException("Nenhum registro encontrado");
+			throw new NotFoundHttpException("Nenhum registro encontrado.");
 		}
 		
 		return $user;
@@ -67,8 +78,10 @@ class PsychologistsController extends Controller
 	 */
 	public function destroy($id)
 	{
+		$this->authorize('destroy', Psychologist::class);
+
 		if (!$this->repository->delete((int) $id)) {
-			throw new NotFoundHttpException("Nenhum registro encontrado");
+			throw new HttpException(Response::HTTP_INTERNAL_SERVER_ERROR, "Não foi possível excluir o registro.");
 		}
 	}
 }
